rename files, functions and improve docstrings
