+++ cg_lims/EPPs/udf/set/base.py
@@ -2,11 +2,11 @@ import click
 from cg_lims.EPPs.udf.set.replace_flow_cell_output_path import replace_flow_cell_output_path
 from cg_lims.EPPs.udf.set.set_barcode import assign_barcode
 from cg_lims.EPPs.udf.set.set_method import method_document
+from cg_lims.EPPs.udf.set.set_novaseq_x_sequencing_settings import set_sequencing_settings
 from cg_lims.EPPs.udf.set.set_ont_sequencing_settings import set_ont_sequencing_settings
 from cg_lims.EPPs.udf.set.set_revio_sequencing_settings import set_revio_sequencing_settings
 from cg_lims.EPPs.udf.set.set_sample_date import set_sample_date
 from cg_lims.EPPs.udf.set.set_samples_reads_missing import set_reads_missing_on_new_samples
-from cg_lims.EPPs.udf.set.set_sequencing_settings import set_sequencing_settings
 from cg_lims.EPPs.udf.set.smrt_cell_metrics import smrt_cell_metrics
 from cg_lims.EPPs.udf.set.smrt_link_run_information import fetch_smrtlink_run_information
 from cg_lims.EPPs.udf.set.updated_sample_volume import updated_sample_volume
