+++ tests/test_status_db_api_client.py
@@ -15,7 +15,7 @@ def test_get_sequencing_metrics_for_flow_cell(
     mocker.patch("requests.get", return_value=mock_sequencing_metrics_get_response)
 
     # WHEN retrieving sequencing metrics for a flow cell
-    result = status_db_api_client.get_sequencing_metrics_for_flow_cell("flow_cell_name")
+    result = status_db_api_client.get_sequencing_metrics_for_illumina_flow_cell("flow_cell_name")
 
     # THEN a list of the parsed sequencing metrics should be returned
     sequencing_metrics = [
